feat(Game): game stories and refactor pt. II
