+++ src/pages/Play/index.tsx
@@ -1,10 +1,30 @@
-import React, { useState } from 'react';
+import React, { useEffect, useRef, useState, KeyboardEvent } from 'react';
 import { useQuery, useMutation } from '@apollo/client';
 import { Game } from './../../components';
 import washington from './../../assets/game_washington.jpg';
 import { QUESTION_QUERY, SAVE_MUTATION } from '../../gql';
 
 const Play = () => {
+  const [submitted, setSubmitted] = useState(false);
+  // const [tip, setTip] = useState<number | null>(null);
+  const handleSubmit = (event: KeyboardEvent<HTMLInputElement>) => {
+    if (event.key === 'Enter') {
+      if (timeoutRef.current) {
+        clearTimeout(timeoutRef.current);
+      }
+      console.log('cum sum!');
+      saveTip({
+        variables: {
+          id: 'lfjsdkfhdkj',
+          tip: Number(event.currentTarget.value),
+        },
+      });
+      // setTip(Number(event.currentTarget.value));
+      setSubmitted(true);
+    }
+  };
+  const timeoutRef = useRef<number>();
+
   const [questions, setQuestions] = useState([
     {
       question: 'Jak velkou má tadydlencten pán hlavu?',
@@ -31,6 +51,18 @@ const Play = () => {
     },
   });
 
+  useEffect(() => {
+    if (questions[questions.length - 1].timeLimit && !submitted) {
+      console.log('started timeout');
+      timeoutRef.current = setTimeout(() => {
+        console.log('ended timeout');
+        if (!submitted) {
+          setSubmitted(true);
+        }
+      }, questions[questions.length - 1].timeLimit * 1000);
+    }
+  }, [questions, submitted]);
+
   if (loading) {
     return <div>loading</div>;
   }
@@ -40,10 +72,8 @@ const Play = () => {
   return (
     <Game
       settings={questions[questions.length - 1]}
-      onSubmit={(value: string) => {
-        console.log('cum sum!');
-        saveTip({ variables: { id: 'lfjsdkfhdkj', tip: Number(value) } });
-      }}
+      isSubmitted={submitted}
+      onSubmit={handleSubmit}
       onFinish={() => {
         if (nextQuestion != null) {
           setQuestions([...questions, nextQuestion]);
